rm model from training file
