+++ algos/maTT/decentralized_ppo_one_network.py
@@ -25,52 +25,13 @@ import numpy as np
 import torch
 import torch.nn as nn
 import torch.optim as optim
-from torch.distributions.categorical import Categorical
 from torch.utils.tensorboard import SummaryWriter
 
-def layer_init(layer, std=np.sqrt(2), bias_const=0.0):
-    torch.nn.init.orthogonal_(layer.weight, std)
-    torch.nn.init.constant_(layer.bias, bias_const)
-    return layer
-
-class PPO(nn.Module):
-    def __init__(self, envs):
-        super().__init__()
-        self.envs = envs
-        key = list(envs.single_observation_space.keys())[0]
-        self.critic = nn.Sequential(
-            layer_init(nn.Linear(np.array(envs.single_observation_space[key].shape).prod(), 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, 1), std=1.0),
-        )
-        self.actor = nn.Sequential(
-            layer_init(nn.Linear(np.array(envs.single_observation_space[key].shape).prod(), 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, 64)),
-            nn.Tanh(),
-            layer_init(nn.Linear(64, envs.single_action_space.n), std=0.01),
-        )
-
-    def get_value(self, x):
-        if isinstance(x, np.ndarray):
-            x = torch.from_numpy(x)
-        x = x.view(-1, x.shape[-2] * x.shape[-1])
-        return self.critic(x)
-
-    def get_action_and_value(self, x, action=None):
-        if isinstance(x, np.ndarray):
-            x = torch.from_numpy(x)
-        x = x.view(-1, x.shape[-2] * x.shape[-1])
-        logits = self.actor(x)
-        probs = Categorical(logits=logits)
-        if action is None:
-            action = probs.sample()
-        return action, probs.log_prob(action), probs.entropy(), self.critic(x)
 
 
-def decentralized_ppo(envs, args, run_name):
+
+
+def decentralized_ppo(envs, model, args, run_name):
     """
     env_fn: 
         lambda: env
@@ -106,7 +67,7 @@ def decentralized_ppo(envs, args, run_name):
     
     assert isinstance(envs.single_action_space, gym.spaces.Discrete), "only discrete action space is supported"
 
-    agent_network = PPO(envs)
+    agent_network = model
     optimizer = optim.Adam(agent_network.parameters(), lr=args.learning_rate, eps=1e-5)
     # agent = Agent(envs).to(device)
     # optimizer = optim.Adam(agent.parameters(), lr=args.learning_rate, eps=1e-5)
@@ -150,7 +111,7 @@ def decentralized_ppo(envs, args, run_name):
     next_obs = envs.reset()
     next_done = torch.zeros(args.num_envs).to(device)
     num_updates = args.total_timesteps // args.batch_size
-
+    ep_length = 0
     for update in tqdm(range(1, num_updates + 1)):
         # Annealing the rate if instructed to do so.
         if args.anneal_lr:
@@ -160,6 +121,7 @@ def decentralized_ppo(envs, args, run_name):
             optimizer.param_groups[0]["lr"] = lrnow
 
         for step in range(0, args.num_steps):
+            #print(f"step {step}") 
             global_step += 1 * args.num_envs
             for i, agent_id in enumerate(next_obs.keys()):
                 obs[step, :, i] = torch.from_numpy(next_obs[agent_id])
@@ -178,18 +140,33 @@ def decentralized_ppo(envs, args, run_name):
             if args.render:
                 envs.envs[0].render()
             # TRY NOT TO MODIFY: execute the game and log data.
+            #import pdb; pdb.set_trace()
             next_obs, reward, done, info = envs.step(action_dict)
+            
             rewards[step] = torch.tensor(reward).to(device).view(-1)
             next_done = torch.Tensor(done).to(device)
-            if "episode" in info.keys():
+            if torch.sum(next_done).item() == args.num_envs:
+                #from IPython import embed; embed()
+                for env_i in range(args.num_envs):
+                    print(f"global_step={global_step}, episodic_return_env={env_i}={torch.sum(rewards[step - ep_length:step, env_i]).item()}")
+                    writer.add_scalar(f"charts/episodic_return_env_{env_i}", torch.sum(rewards[step - ep_length:step, env_i]).item(), global_step)
+                writer.add_scalar("charts/episodic_length", ep_length, global_step)
+                ep_length = 0
+                envs.reset()
+                break
+            """if "episode" in info.keys():# and torch.sum(next_done).item() == args.num_envs:
+                print(info)
+                print(next_done)
+                from IPython import embed; embed()
                 assert torch.sum(next_done).item() == info["episode"].shape[0]
                 for env in range(args.num_envs):
                     print(f"global_step={global_step}, episodic_return={info['episode'][env]['r']}")
                     writer.add_scalar("charts/episodic_return", info["episode"][env]["r"], global_step)
                     writer.add_scalar("charts/episodic_length", info["episode"][env]["l"], global_step)
-                break
+                break"""
+            ep_length += 1
         
-        
+        #print(f"finished stepping {args.num_steps} in env")
         # bootstrap value if not done
         advantages = torch.zeros((args.num_steps, args.num_envs, args.nb_agents))
         returns = torch.zeros((args.num_steps, args.num_envs, args.nb_agents))
